rename diagnostics.offline.compute_diags to .compute
