+++ workflows/argo/offline-diags.yaml
@@ -46,7 +46,7 @@ spec:
             {{inputs.parameters.validation_data_config}}
             EOF
     
-            python -m fv3net.diagnostics.offline.compute_diags \
+            python -m fv3net.diagnostics.offline.compute \
               {{inputs.parameters.ml-model}} \
               validation_data.yaml \
               {{inputs.parameters.offline-diags-output}}
